feat(ready): unified §5.5 ready predicate; typed fields in ready JSON (beads_rust-ready-predicate-9n4c)
Predicate (build_ready_issue_candidates_query, all callers incl.
scheduler/coordination): verify Some+non-empty+single-line; priority<=2
requires >=1 principles citation with non-empty name/decision (JSON1
fail-closed on malformed); wave gate holds wave n while any lower-wave
issue is open/in_progress. Condition 2 (#432): missing local blockers
ignored via INNER JOIN in load_direct_blockers_impl; closed/tombstone
already ignored; other statuses keep #357 semantics.

br ready --json/toon rows now carry verify, principles, wave, pin
(§5.5 shape) via extended Command projection + ready_issue_from_row.

Fixtures: shared builders default to dispatchable beads (verify +
principles citation); §8.4 named tests added: ready_omits_missing_verify,
ready_omits_p1_missing_principles, ready_holds_wave_2_while_wave_1_open,
ready_ignores_closed_blocker (+ e2e §5.5 shape test). Scenario
dependency-story expectation updated: CLI-created beads without typed
verify are no longer dispatchable until fence-import stamps them.

Also folds in the captain-assigned clippy fix: e2e_schema_migration_upgrade
TARGET_SCHEMA_VERSION becomes u32 (no cast_possible_truncation), plus
fmt drift fix in update.rs (HEAD-owned from abe91f30).

Gates: lib 2744/0/8; storage_ready 182/0; schema-upgrade 161/0;
e2e_ready 188 pass + 3 pre-existing HEAD failures (rework-status policy /
external-deps trio, fail at pristine HEAD too - reported not chased);
clippy --all-targets -D warnings exit 0; fmt --check clean.

PRINCIPLES:
encode-lessons-in-structure - one predicate in SQL; flywheel can delete
parse_verify_fence and trust br ready --json.
prove-it-works - each exclusion rule has a named test.

+++ tests/e2e_schema_migration_upgrade.rs
@@ -38,7 +38,7 @@ fn fixture_dir() -> PathBuf {
 /// Migration target asserted by this suite: `CURRENT_SCHEMA_VERSION` (18, the
 /// ADR-0001 typed work-ledger schema). Kept as a named constant so a future
 /// bump is a one-line change plus fixture refresh.
-const TARGET_SCHEMA_VERSION: u64 = 18;
+const TARGET_SCHEMA_VERSION: u32 = 18;
 
 fn install_fixture_workspace(workspace: &BrWorkspace, db_gz: &str, issues: &str, config: &str) {
     let beads_dir = workspace.root.join(".beads");
@@ -140,7 +140,7 @@ fn upgrade_fixture_end_to_end(
     assert_eq!(plan_json["from_version"].as_u64(), Some(expected_from));
     assert_eq!(
         plan_json["to_version"].as_u64(),
-        Some(TARGET_SCHEMA_VERSION)
+        Some(u64::from(TARGET_SCHEMA_VERSION))
     );
     let plan_token = plan_json["plan_token"]
         .as_str()
@@ -173,7 +173,7 @@ fn upgrade_fixture_end_to_end(
     let run_id = applied_json["run_id"].as_str().expect("run id").to_string();
     assert_eq!(
         header_user_version(&db_path),
-        TARGET_SCHEMA_VERSION as u32,
+        TARGET_SCHEMA_VERSION,
         "{label}: post-apply schema"
     );
     for table in [
@@ -272,7 +272,7 @@ fn upgrade_fixture_end_to_end(
     );
     assert_eq!(
         header_user_version(&db_path),
-        TARGET_SCHEMA_VERSION as u32,
+        TARGET_SCHEMA_VERSION,
         "{label}: rejected stale apply must not mutate the database"
     );
 
@@ -349,7 +349,7 @@ fn upgrade_fixture_end_to_end(
     );
     assert_eq!(
         header_user_version(&db_path),
-        TARGET_SCHEMA_VERSION as u32,
+        TARGET_SCHEMA_VERSION,
         "{label}: re-apply after undo"
     );
 
@@ -381,7 +381,7 @@ fn upgrade_fixture_end_to_end(
     );
     assert_eq!(
         header_user_version(&db_path),
-        TARGET_SCHEMA_VERSION as u32,
+        TARGET_SCHEMA_VERSION,
         "{label}: post-migration write must not move the schema version"
     );
 }
